requirements.txt
# Budget Amico - Requirements
# Framework GUI
flet>=0.28.0
flet-desktop>=0.28.0

# Database
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0

# Data Processing & Export
pandas>=2.3.0
openpyxl>=3.1.0
numpy>=2.3.0

# Utilities
python-dateutil>=2.9.0
requests>=2.32.0
yfinance>=0.2.40

# Build Tools (optional, for creating executables)
pyinstaller>=6.16.0

# Security
cryptography>=42.0.0
